Fix: Match short namespace symbols
When using short namespaces instead of full namespaces, the symbol
should properly match in order to detect usages of a required package.

Example:

    A package with PSR-4 namespace "Cocur\Slugify\" is required. The
    project code base contains the following class:

    <?php

    use Cocur\Slugify;

    final class MyBeautifulClass
    {
        public function __construct(
            private readonly Slugify\Slugify $slugify,
        ) {
        }
    }

In this example, the namespace symbol would not match, because the
registered PSR-4 namespace contains a trailing backslash whereas
the imported namespace in the class does not.

To resolve incorrect behavior in such cases, a trailing backslash is
now always appended to the constructed namespace symbols.