Handle name resolving with PHP-Parser's NameResolver
Name resolving of namespaced symbols (such as classes,
functions and constants) is now done through PHP-Parser's
NameResolver. By using it as an additional node visitor,
we add a safe additional layer on top of the collection
strategies provided by the symbol-parser package.

This also makes the fix introduced with 13fd5e9 obsolete
now, as the NameResolver is more accurate.